Messed up Mailer (Introduced by 4.1 -> 4.2)
In 4.1 it was possible returning a count of sent emails from Mail::send().
In 4.2 you'll always get null returned. (Because Mail::send() is now returning the void method `sendSwiftMessage()`)

(`sendSwiftMessage()` was returning the count in  4.1 ...)


If that was intended the send method needs an update which I'm providing with this pull request.
Anyways I hope it was not and the send() method will return the count of sent emails again in future. (Was using this feature and liked it alot)