Replace Application contract with Container
This is a follow-up for https://github.com/laravel/framework/pull/29619

Mior correctly noted that Managers only need to have a container instance to function properly. The fact that the default container we inject here is an instance of the Laravel application is an implementation detail.

This PR contains no breaking changes other than renaming the $app property to $container. The new type-hint still allows for the current Application instance to be passed through but makes it clear that this could be any implementation of a container.

I've also made $config a dedicated property on the Manager class as it's widely used on our own adapters. This makes it easy to reference it. The container array calls were replaced by `make` methods because `ArrayAccess` isn't part of the Container contract (and it shouldn't be).